# docs_src/openapi_thumbnail.json
{"openapi": "3.0.2", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/": {"get": {"summary": "Root Handler", "operationId": "root_handler__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/healthz": {"get": {"summary": "Healthz Handler", "operationId": "healthz_handler_healthz_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/topz": {"get": {"summary": "Topz Handler", "operationId": "topz_handler_topz_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/thumbnail/{paper_uuid}": {"post": {"summary": "Create Thumbnail", "operationId": "create_thumbnail_thumbnail__paper_uuid__post", "parameters": [{"required": true, "schema": {"title": "Paper Uuid", "type": "string", "format": "uuid"}, "name": "paper_uuid", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/thumbnail/{paper_uuid}/{image_id}": {"get": {"summary": "Read Thumbnail", "operationId": "read_thumbnail_thumbnail__paper_uuid___image_id__get", "parameters": [{"required": true, "schema": {"title": "Paper Uuid", "type": "string", "format": "uuid"}, "name": "paper_uuid", "in": "path"}, {"required": true, "schema": {"title": "Image Id", "type": "string"}, "name": "image_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"type": "string"}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}